ai-mammography/app/requirements.txt
fastapi
uvicorn
python-multipart
keras
tensorflow
joblib
keras
matplotlib
numpy
opencv_python
Pillow
pydicom
scikit_learn
scikit-image
torch
torchvision
python-dotenv
pydantic-settings
email-validator